Add workaround for ufmt related codegen bug
As reported by @peacememories in #1, the hint::unreachable_unchecked
intrinsic is leading to a codegen bug if the ufmt usize formatter
is not inlined.

This bug has to be fixed in the compiler and was reported upstream
as avr-rust/rust#148.

Until the actual cause and a solution are found, this fork of ufmt
which does not use the above mentioned intrinsic, should be used as
a workaround.

Signed-off-by: Rahix <rahix@rahix.de>