generic: serial: Refactor impl macro by using paste
Instead of forcing hals to spell out the names for each register and
field, use paste [1] to produce the names on the fly.  This makes the
macro invocation considerably less complex and allows easier addition of
new features as access to new fields no longer requires changes across
all HALs.

[1]: https://crates.io/crates/paste

Signed-off-by: Rahix <rahix@rahix.de>