Support for open-drain mode
Solution for issue #16.

A new port mode (OpenDrain) that Suports open-drain mode with external pull-up.
The solution uses the DDRxn to switch between a high state with the pin
input floating mode where the external pull-up is setting the pin to high state,
and low state by setting the pin in output low to pull the pin low.

Input traits is reading the external electrical state independent of if
the pin is set high or low.

Signed-off-by: Karl Thorén <karl.h.thoren@gmail.com>