Rename OpenDrain to TriState and update comments
Commit e00abb6c introduced a Mode that uses DDRx to toggle the pin
state between driving low and floating high (presumably using an
external pull-up).  This commit renames that mode to TriState.

Calling this mode OpenDrain suggests that it would be reasonable for
the external pull-up to connect to any rail above GND, but this is not
generally valid for the AVR I/O cell.  This output can be safely
pulled-up only as far as VCC + 0.5 before the protection diode goes
into conducting mode.