broken: rename  to  in impl_twi_i2c macro
Error:

    error[E0422]: cannot find struct, variant or union type `I2c` in this scope
      --> chips/atmega328p-hal/src/lib.rs:31:20
       |
    31 |         pub struct I2c {
       |                    ^^^ not found in this scope

I don't understand something. I thought that the chips were defining the
`I2c` struct and then passing it into the macro for expansion. The macro
would then match against the struct elements, and then expand them in
the macro expansion. If the chips defined the I2c struct, then this
error would not be here. So where is the I2c struct defined? Here? Or in
the macro?