treewide: i2c: Rename I2c struct to I2cMaster
In preparation for a future addition of an I2C slave driver, rename the
I2C master driver to `I2cMaster`.  To ease migration, add a type-alias
for the old name which is marked as deprecated.  This alias will be
removed at a later time when most users have switched to the new name.

Co-authored-by: toddg <toddg@zwrob.com>
Signed-off-by: Rahix <rahix@rahix.de>