generic: usart: Make macro usage internal only
It is showing more and more that generating the entire API for the HAL
crates from macros is a bad idea.  Reasons include:

 - It obscures the code.  People have trouble finding where a certain
   type is defined, especially if they are not too well aquainted with
   rust macro_rules!().

 - It breaks auto-completion in IDEs because they'd need to introspect
   the macros.

 - It leaves many types with the same methods instead of a single one.
   This also makes code that is generic over the exact peripheral
   impossible to write.

 - It makes it much harder to cleanly support different peripherals
   under a common interface.

To mitigate these problems, shift to a different design:  A single
`Usart` type provides the public API and is generic over the USART
peripherals which all implement a common interface called `UsartOps`.
Only this trait implementation will be macro generated and thus there is
no user-facing macro-generated API left.

This design also makes it much easier to implement support for multiple
different hardware peripherals: As long as the `UsartOps` are expressive
enough to fit all of them, we can simply provide multiple macros for
generating the implementation for each one.