generic: port: Add new port implementation
Replace the existing port implementation with a new design.  With the
new design, there is a single `Pin` type for representing pins and
specific pins can be discerned via one of its generic parameters.

One one hand, this eases visibility of the public API as it is no longer
hidden inside a magic macro and on the other hand it allows to easily
support the different IO cells of old and new AVR MCUs.

Please refer to the included documentation in this commit for more
details.

Ref: #114