atmega2560: adc: Use custom MUX enum instead of relying on PAC
The previous code was based on the incorrect PAC enum which contained
enumerated values which actually do not fit the field they were
specified for.  As those were removed in avr-device 0.3, this
implmentation here also no longer compiles.

Instead, use a similar approach to the atmega32u4-hal where a custom
`AdcMux` enum is defined which we use to represent all supported MUX
values in a compact way.