generic: usart: Make impl_usart_traditional!() work with new pins
Adjust the `impl_usart_traditional!()` macro to support the new design
of pin types.

This leads to an interesting issue: Because with the new design, neither
the type (e.g. `USART0`) nor the trait (`UsartOps`) nor the pin type are
from the HAL crate, we run into the orphan rule.

Mitigate this by expecting a newtype wrapper around the peripheral, and
as such do the register accesses through a `.0` accessor.  In the future
we might want to make this accessor configurable.