atmega-hal: Drop custom `Peripherals` struct
It seems the approach of defining a custom `Peripherals` struct does not
scale well: We'd need to replicate all the fields from the respective
MCUs `Peripherals` struct to provide a usable API.  But doing so would
be a huge effort.

Instead, look for a design where the custom `Peripherals` is not needed.

As a first step, convert the `Pins` initialization to a macro-based
approach.  A `pins!()` macro is provided which moves the relevant fields
out of the `Peripherals` struct.  This must be a macro because it must
happen as part of the `main()` function where the `Peripherals` instance
can linger in its partially moved state.