generic: usart: Introduce a type-parameter for orphan circumvention
Because HAL crates neither own the `UsartOps` trait nor the `USARTn`
instance, they are not allowed to implement it for the type.  However,
the orphan rules allow us to deal with this by adding a generic
parameter to `UsartOps` which is local to the HAL crate.

The new `H` parameter serves exactly this purpose and nothing more.
A HAL should define a type and use it for `H` in all HAL
implementations.  Ideally, it should also export type-aliases for all
types from `avr-hal-generic` with the `H` parameter already substituted.