generic: port: Add "input mode forgetting"
Having the input mode encoded in the type is useful for static
guarantees but cumbersome in situations where one does not care.

Provide a way to "erase" the type-information about input mode, by way
of a `.forget_imode()` method.  The resulting type behaves just like
a normal input pin except that pull-up and floating pins have the same
type.