Introduce the "ravedude"
ravedude is a wrapper around avrdude to make Rust development for AVR
microcontrollers as easy as possible.  ravedude is meant to replace the
current hackish runner scripts with something that works out of the box.

In addition to managing the avrdude invocation, ravedude also comes with
a super basic terminal emulator to make accessing the serial console
easy.  In a similar spirit to `probe-run`, this should provide an
embedded development experience which is as close to "desktop
development" as possible.