generic: adc: Refactor ADC HAL
Reimplement the ADC HAL with a design similar to the other new
implementations.  This means we have a common `Adc<...>` type defined in
avr-hal-generic which internally uses a concrete ADC peripheral which
implements the `AdcOps` trait.

Additionally, an `AdcChannel` trait is used to mark channels which can
be used with a certain ADC.