examples: uno: hc-sr04: Handle panic due to overflow
Some HC-SR04 sensor builds keep their echo pin in high state for long
time, instead of keeping it in low state when no object is detected.
When this happens the multiplication of the TIMER1 value with 4 goes
into overflow, resulting panic, crashing the program execution.

Fix this issue by handling the saturating u16 value the same way as if
the echo pin never had gone high.

This commit is a forward-port of 885e8ec6d6d2 ("uno: hc-sr04: Handle
panic due to overflow") to the new version of avr-hal.