Add autoreset logic for arduino leonardo.
The arduino leonardo uses a funny method for autoreset. Whenever
you connect to the usb driven serial port at 1200 baud, it resets.
I have implemented this logic. If it can't find a device to
connect to, then it will fall back to asking you to manually
reset the device.

This protocol is not yet implemented in avr-hal, so this will
only work with loading firmware on a device that was flashed with
the arduino tools. Until avr-hal implements this protocol for
the leonardo, you will have to continue to manually reset if
during programming. However, I think that it is useful to
implement this side of the protocol as a motivation for getting
the USB CDC serial device working on the leonardo.

I tested it numerous times, and it seems to be reliable when I
flash a device previously flashed with the arduino tools.