ravedude: Add autoreset logic for Arduino Leonardo
The arduino leonardo uses a funny method for autoreset. Whenever you
connect to the usb-serial port at 1200 baud, it resets.  Implement this
logic.  If it can't find a device to connect to, then it will fall back to
asking you to manually reset the device.

This protocol is not yet implemented in avr-hal, so this will only work
with loading firmware on a device that was flashed with the Arduino
tools.  Until avr-hal implements this protocol for the Leonardo, you will
have to continue to manually reset it during programming.