examples: Fix build-std settings for compiler-builtins
This is a workaround to ensure intrinsics get linked in from avr-libc
instead of Rust's compiler-builtins.  The problem is that the versions
from compiler-builtins have slightly different calling convention and as
such cannot actually be substituted for the versions from avr-libc right
now.

This should fix a lot of the observed problems with 32-bit arithmetic.

Ref: https://github.com/rust-lang/rust/issues/82242#issuecomment-997400373