Examples: Added proper sleep mode control register config to uno-timer.rs example
Previously in the example the sleep mode control register was never configured for sleep mode.
To do this, the sleep enable (SE) bit has to be set active. Due to the examples layout, it was
not apparent that the controller didn't go to sleep, but it is easily proven by putting any kind
of visible functionality inside the loop.