Add support for ATmega8
ATmega8 is more challenging in some ways, because it does not work like
the atmegas that were already implemented

  1. EEPROM does not have modes (erase, erase and write), completely
     different mechanism is required.
  2. WDT registers have different names "s" is missing from wdtcsr.
  3. Does not have DIDR for ADCs
  4. USART does not have "0" at end. It's just USART, not USART0. It has
     registers UBRRL, UBRRH instead of just one UBRR. I don't think it's
     possible to merge into just one register using patches as the UBRRH
     is shared with UCSRC and it's not directly next to UBRRL.
  5. WDT does not have the same prescalers. It also does not support 4k
     ms and 8k ms. The rest of the prescalers does not exactly match the
     values.

For the time being, the USART problem is solved with a custom UsartOps
implementation in `atmega-hal`.  In the future, this should be revisited
and a better solution needs to be found.