ci: Make ravedude job work again
There is a problem with actions-rust-lang/setup-rust-toolchain where the
toolchain specified in the root rust-toolchain.toml file always takes
precedence over everything else.  This means ravedude is currently built
with the nightly toolchain instead of stable as we intend.

As a workaround for the time being, make ravedude build with the nightly
toolchain.