add `embedded-hal=1.0.0` dependency
this does not yet implement any 1.0.0 APIs, this is just a first step to
make them available for implementation.

the existing v0.2 dependency is being kept but is internally renamed to
`embedded-hal-v0`. the re-export as `avr_hal_generic::hal` still points
to v0.

v1 APIs can now be implemented step-by-step. once all APIs are
implemented the v0 APIs can be removed as a breaking change.