implement `embedded-hal` v1 `DelayUs`
the implementation for nanoseconds will only do so with
microsecond-accuracy as there's currently no implementation for ns.
this is ok(-ish), as the trait allows to delay for longer than
specified. if an implementation with us-accuracy can be done this can be
done so later as a follow-up.

the ns-implementation is done explicitly as the blanket implementation
would otherwise transform it into us-calls which then lose precision.
the ms-implementation was left to the blanket-implementation as it
correctly steps it down to ns, as is also done for the old v0
implementations.