add uno blinking example using only `embedded-hal` traits
this is the same as `uno-blink.rs`, however the actual blinking is done
by a function which does not make any use of any `avr-hal` specific
APIs, instead it solely relies on `embedded-hal` v1 traits.

note that the current v1.0.0-rc.3 version of `embedded-hal` still has
`toggle` as a separate trait fn on `ToggleableOutputPin`, however that
has been removed and v1.0.0 will ship with a blanket impl of `toggle` in
`StatefulOutputPin`. so when updating to v1.0.0 the manual
`set_low`/`set_high` calls can be removed, making the implementation
equivalent to `uno-blink.rs`.