migrate `extern crate` to `pub use`
`extern crate` is from the old rust 2015 edition. since the 2018 edition
`extern crate` is no longer needed. here it was used to re-export
certain crates (through `pub extern crate`) which are being referenced
mainly in macros (which is why e.g. `paste` is re-exported). this can
also be achieved using `pub use`.

`avr-hal-generic` internally now has to use `crate::hal` in some cases
where it could previously just use `hal` as the renaming is now bound to
the usage in the crate and not to the imported crate. in most cases i've
just replaced `hal::` usage with `embedded_hal_v0::` usage since it
makes it more obvious which e-h version it is about (and in the future
`hal` should either be removed or be replaced with e-h v1.0).

for external consumers nothing changes.