remove `ufmt` re-export from `avr-hal-generic`
note: i've explicitly dropped the re-export of `ufmt` as it's not used
in this repository* and i don't see `ufmt` as an integral part of these
crates. consumers can (and should) just add it as their own dependency.
this is a breaking change for consumers which currently do something
like `use avr_hal_generic::ufmt`.

\* clarification: `ufmt` itself is being used in some places, but the
   re-export isn't.