remove `void` dep, use `core::convert::Infallible` instead
there's no need for the `void` dependency anymore. accordingly it can be
removed. this will also help with implementing `embedded-hal` v1 traits
as some of them rely on (or at least have better support for)
`core::convert::Infallible`.

this is a breaking change for consumers as they have to e.g. replace
`void_unwrap()` calls with `unwrap()`.