implement `embedded-io` v0.6 `Read` & `Write`
with `embedded-hal` v1 the USART traits have been removed in favour of
the new `embedded-io` crate.

this adds a (very basic) implementation for `Read` and `Write`. other
traits (such as the `*Ready` or `BufRead` traits) have not (yet) been
implemented and some (like `Seek`) probably can't be implemented for
this HAL.

a better implementation might use a buffer in the background to receive
more than one byte at once.

see also #249 for a related PR.

this is part of #468