examples: nano168: Drop panic example
This example is completely identical to the one for Arduino Uno so let's
just keep that one as the reference.

The motivation for doing this now is a weird CI failure where this panic
example seems to overflow flash storage space [1].  This is not
reproducible on a local build.  So let's just remove the example and
pray that's the end of it.....

[1]: https://github.com/Rahix/avr-hal/actions/runs/7493773898/job/20400220985#step:5:58