Use `unwrap-infallible`
After removing `void`, we're still missing a way to explicitly unwrap
infallible results such that it is clear that no panic can occur in this
location.  Because the standard library does not yet provide a solution
to this, let's use the `unwrap-infallible` crate [1] in the meantime.

This topic was also discussed in more detail in the embedded-hal project [2].

[1]: https://crates.io/crates/unwrap-infallible
[2]: https://github.com/rust-embedded/embedded-hal/issues/329