ravedude: Use a different avrdude config for avrdude 7.X
The config we've been shipping so far was only for avrdude version 6.X.
This config does not work when being used with avrdude 7.X versions.

As we do not have control over the installed avrdude version at this
time, we need to provide different configurations depending on the
avrdude version we find.

As a stop-gap solution, import the released default config for avrdude
7.3 and attempt using it for all avrdude 7.X versions.

In the future, we'll have to rethink this.  It is probably a much better
solution to simply vendor avrdude entirely.