generic: spi: Add escape hatch for the ChipSelectPin
There are situations where the `ChipSelectPin` abstraction is not enough -
for example when a user wants to use the `ChipSelectPin` as a PWM pin.

To enable such usecases anyway, add an escape hatch for getting back the
original `Pin<>` type that's hidden in the `ChipSelectPin`.  This is
unsafe, for obvious reasons.