ravedude: Add support for Mighty Core ATMEGA1284P boards
This adds support for the Arduino DIP-40 Development Board [1] with an
ATMEGA1284P using the Mighty Core Arduino core and bootloader, or really
any board using a ATEMEGA1284P with the Mighty Core Arduino core
installed.

Tested and verified to work with a ATEMEGA1284P. Theoretically variants
of this can be made for other ATMEGA DIP40 MCUs, but I am not in a
position to test that.

[1]: https://www.tindie.com/products/MCUdude/dip-40-arduino-compatible-development-board/