Fix pin toggle for some old microcontrollers
Some older AVR microcontrollers do not support the pin-toggling via
write to the `PINx` register.  For these, we have to manually implement
toggling via a read-modify-write sequence.

Augment the `avr-hal-generic` macro to support this manual toggle
implementation and use this for all MCUs which do not support the
new-style toggling.

Additionally, the read-modify-write sequence has to be wrapped in a
critical section to ensure the updates are atomic.