generic: Accommodate for new return value of write() and modify()
In svd2rust 0.34.0, the `.write()` and `.modify()` register access
functions were changed to return the written register value.  As
avr-device is upgrading to a more recent svd2rust version, we need to
adjust accordingly.

Fixup the few places where the return value of `.write()` or `.modify()`
mattered due to expression form calls.