ravedude: update and revise flake
Dependency locks were updated, and the version of Rust used to build is
now synchronized to the `ravedude/rust-toolchain.toml` file. The precise
date of the latest stable was added to this file, and it should be
updated when we raise our requirements. This way Nix will break
explicitly (since it won't have versions from the current future),
instead of silently using the old locked value.