Provide a direnv-able nix flake shell at the root
This makes it easier for Nix users to develop the package. It brings-in
the recommended Rust infrastructure as well as necessary runtime
programs. It does not actually build the project, so contributors are
still expected to either build as a library or from one of the examples
to test for errors. This should be done manually.