ravedude: Add numeric output modes
Sometimes printing ascii isn't useful, and the raw hex, binary or
decimal data should be printed instead.

Add a new `output-mode` setting which can be set to `ascii`, `hex`,
`dec`, or `bin`.  In the numeric modes, one of two newline strategies
can be selected: `newline-after` will start a new line after N bytes and
`newline-on` will start a new line after a specific byte is received.
All these options are placed in the `[general]` section of
`Ravedude.toml`.

Finally, a `-o`/`--output-mode` commandline flag can be used to choose a
different output-mode on the fly.

Commit-message-by: @rahix